+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#improved)
     * Improved flex router event to include directory
 3. [](#bugfix)
+    * Fixed caching issues in dynamic flex forms
     * Fixed flex content in unauthorized module causing the whole page to become unauthorized
 
 # v1.1.9